requirements.txt
discord.py
urlextract
